Add socketsink.py for sinking data provided by the Saleae HLA
This is a basic script intended to showcase the functionality of the
HLA and minimizing the amount of effort one has to put in to working
with sockets if all they want to do is get some data out.

socketsink.py simply creates a configurable listener socket that watches
for data provided by the SaleaeSocketSource HLA and prints that data to
STDIN. There is some basic error handling mean to recover from the most
common errors I have encountered so far, but this is by no means robust.

You could use this script on its own and pipe the data to another
program or file for further processing. However, for more robust error
handling and faster processing it would be best to write your own socket
listener.

This was primarily designed with the intent of redirecting data for use on
the same machine your Saleae is connected to (as the defaults are
configured to do), but given that the underlying transport is network
ready sockets, I have confirmed that this also works as expected between
machines (assuming no firewalls or other blockers get in the way).