Implement bi-directional sockets
This commit adds support for client and server to communicate data in both
directions, meaning not only can we get data out of Saleae, we can get it back
in as well.

This still needs some work on the Analyzer side to add the ability to parse the
incoming data and transform it back into an AnalyzerFrame, but I'm hoping
that part is comparatively simple.

It's also probably time to do some renaming since the source/sink
terminology no longer really applies.