Important improvements to timing, and added some decode examples
This commit fixes an important issue with the timestamp format I was using for
the frame data. I was converting this to a datetime.datetime object then
serializing that to ISO formatted dates. But datetime.datetime does not
have nano-second resolution so I was losing timing information and the data
was coming back in slightly off. I managed to figure out how to serialize
and then recreate a SaleaeTime object directly so now I'm just using that
representation.

This commit also adds support for the 'text' data type, which should be good
for several general use cases, though I'm not really happy with how it shows
up in the data table. You change this representation through the
result_types dictionary, so I looked into adding a way to dynamically extend these
definitions. Unfortunately it looks like this is consumed by some higher level
code in Saleae probably before the class is even instantiated and then it
isn't parsed again so any changes after that have no effect.

I need to give some more thought to what kind of result_types I want to
natively support, but this is getting very close to a workable final
implementation.