Implement dynamic loading of response handlers
This commit implements the final missing piece of the client side
handling code, which is the ability for users to easily implement
their own response handlers and specify their execution from the
command line. This is implemented through the `-r/--responder`
option.

The README is also updated with some introductory text on how
to implement your own responder.