More README.md updates
Mostly tweaks to readability and spelling. Moved the "Usage of socketclient.py"
section up higher in the document as I think it is more interesting to the average
user, and anyone looking for the Socket Data Transport details will be willing to
read however much of the README is necessary.